+++ library/FieldUtils/OutputModules/OutputFld.cpp
@@ -39,7 +39,7 @@ using namespace std;
 #include <boost/core/ignore_unused.hpp>
 #include <boost/format.hpp>
 
-#include <LibUtilities/BasicUtils/FileSystem.h>
+#include <LibUtilities/BasicUtils/Filesystem.hpp>
 
 #include "OutputFld.h"