+++ library/FieldUtils/OutputModules/OutputVtkBase.cpp
@@ -40,7 +40,7 @@ using namespace std;
 #include <boost/core/ignore_unused.hpp>
 #include <boost/format.hpp>
 
-#include <LibUtilities/BasicUtils/FileSystem.h>
+#include <LibUtilities/BasicUtils/Filesystem.hpp>
 
 #include "OutputVtkBase.h"