+++ solvers/LinearElasticSolver/EquationSystems/IterativeElasticSystem.cpp
@@ -35,7 +35,7 @@
 #include <boost/format.hpp>
 
 #include <GlobalMapping/Deform.h>
-#include <LibUtilities/BasicUtils/FileSystem.h>
+#include <LibUtilities/BasicUtils/Filesystem.hpp>
 #include <LibUtilities/Foundations/Interp.h>
 #include <LocalRegions/MatrixKey.h>
 #include <MultiRegions/ContField.h>