+++ tests/MetricFileExists.cpp
@@ -40,7 +40,7 @@
 
 #include <MetricFileExists.h>
 
-using namespace boost::filesystem;
+#include <LibUtilities/BasicUtils/Filesystem.hpp>
 
 namespace Nektar
 {
@@ -85,7 +85,7 @@ bool MetricFileExists::v_Test(std::istream &pStdout, std::istream &pStderr)
     boost::ignore_unused(pStdout, pStderr);
 
     bool success = true;
-    auto pwd     = boost::filesystem::current_path();
+    auto pwd     = fs::current_path();
 
     // Check each pattern in turn
     for (auto it = m_fileCounts.begin(); it != m_fileCounts.end(); ++it)
@@ -95,7 +95,7 @@ bool MetricFileExists::v_Test(std::istream &pStdout, std::istream &pStderr)
 
         // Examine each file in the current path and check if it matches the
         // pattern provided. Count the number of files which match.
-        for (auto &e : boost::make_iterator_range(directory_iterator(pwd), {}))
+        for (auto &e : fs::directory_iterator(pwd))
         {
             boost::cmatch matches;
             if (boost::regex_match(e.path().string().c_str(), matches, r))
@@ -125,13 +125,13 @@ void MetricFileExists::v_Generate(std::istream &pStdout, std::istream &pStderr)
     boost::ignore_unused(pStdout, pStderr);
 
     // Update File counts.
-    auto pwd = boost::filesystem::current_path();
+    auto pwd = fs::current_path();
 
     for (auto it = m_fileCounts.begin(); it != m_fileCounts.end(); ++it)
     {
         int cnt = 0;
         boost::regex r(it->first.c_str());
-        for (auto &e : boost::make_iterator_range(directory_iterator(pwd), {}))
+        for (auto &e : fs::directory_iterator(pwd))
         {
             boost::cmatch matches;
             if (boost::regex_match(e.path().string().c_str(), matches, r))