+++ CHANGELOG.md
@@ -20,6 +20,7 @@ v5.5.0
 - Replace `boost::random` with `std::random` (!1673)
 - Add Robin BC to xxt full solver (!1679)
 - Update Parareal file output and tidy-up Parareal/PFASST interpolation (!1678)
+- Move from `boost::filesystem` to `std::filesystem` (!1674)
 
 **ADRSolver**
 - Add support for spatially-constant, but variable direction, diffusion to